create get_random_daily_puzzle module
